Move `KeySelect` to `AttributeKey`
